Resolve /broker/XXX and /user/XXX in output
The circonus json output contains references to other items, such as brokers
and users, and displays them as /broker/123. When circonusvi encounters these,
it resolves them in the api, and displays a human readable description of the
resource in a "comment". Json doesn't support comments, so we just add a line
containing '# foo' before any matching line after json encoding, and strip out
those lines before decoding.

Any API calls to resolve the resources are cached, and an api call is only
ever made if a given endpoint is unknown (e.g. the first run, and if you have
a new enterprise broker). If you rename a broker, and the old name is cached,
you can remove the cache file to start over. Later on an option may be added
to clear the cache.