Use simplejson for python 2.6
Currently circonusvi uses the object_pairs_hook feature of the json parser for
dealing with adding new resources (you can end up with duplicate keys). This
feature isn't available in Python 2.6 and you get an error, but it is
available with simplejson. If simplejson is available, we use that instead of
the built in python json module.